Extract draft cleanup into shared helper function
